Local Agent/dataprocessing/florence_embedding/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
open-clip-torch>=2.20.0
tqdm>=4.65.0






